backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
pydantic==2.7.1
pyyaml==6.0.1
python-dotenv==1.0.1
tenacity==8.5.0
gunicorn==22.0.0